src/python/scrappers/pyproject.toml
[tool.poetry]
name = "scrappers"
version = "0.1.0"
description = "Scrappers for collecting data from supermarkets in Sweden."
authors = ["Erik Dao <cuong@neuraltalks.io>"]
packages = [{include = "scrappers", from = "src"}]
classifiers = [
    "Programming Language :: Python :: 3.9",
]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.10.0"
python-dotenv = "^0.21.0"
tornado = "^6.2"
pymongo = "^4.3.3"
motor = "^3.1.1"
loguru = "^0.6.0"
sentry-sdk = "^1.15.0"
click = "^8.0.3"
luigi = "^3.2.1"

[tool.semantic_release]
branch = "main"
version_variable = "scrappers/__init__.py:__version__"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
